removed waste code
