--- core/templates/pages/exploration-editor-page/editor-tab/templates/modal-templates/customize-interaction-modal.template.html
@@ -1,138 +1,0 @@
-<div class="oppia-interaction-header-container">
-  <div class="modal-header oppia-modal-header" ng-class="{'modal-header-blue': StateInteractionIdService.displayed && (customizationModalReopened || hasCustomizationArgs)}">
-    <h3 ng-if="!StateInteractionIdService.displayed">
-      Choose Interaction
-    </h3>
-
-    <h3 ng-if="StateInteractionIdService.displayed && (customizationModalReopened || hasCustomizationArgs)">
-      Customize Interaction (<[INTERACTION_SPECS[StateInteractionIdService.displayed].name]>)
-    </h3>
-    <i class="fa fa-times" ng-click="cancel()"></i>
-  </div>
-</div>
-
-<div class="customize-interaction-body-container modal-body">
-  <div ng-if="!StateInteractionIdService.displayed" ng-cloak>
-    <ngb-tabset>
-      <ngb-tab ng-repeat="category in ALLOWED_INTERACTION_CATEGORIES"
-               heading="<[category.name]>" active="category.active"
-               class="protractor-test-interaction-tab-<[getHyphenatedLowercaseCategoryName(category.name)]>">
-        <br>
-        <div ng-repeat="interactionId in category.interaction_ids"
-             ng-click="onChangeInteractionId(interactionId)"
-             class="oppia-interaction-tile protractor-test-interaction-tile-<[interactionId]>"
-             title="<[INTERACTION_SPECS[interactionId].description]>">
-          <img ng-src="<[getInteractionThumbnailImageUrl(interactionId)]>">
-          <div class="oppia-interaction-tile-name fx-row fx-main-center fx-cross-center">
-            <[INTERACTION_SPECS[interactionId].name]>
-          </div>
-        </div>
-      </uib-tab>
-    </uib-tabset>
-  </div>
-
-  <div ng-if="StateInteractionIdService.displayed && hasCustomizationArgs" class="interaction-editor-form protractor-test-interaction-editor">
-    <div ng-repeat="customizationArgSpec in customizationArgSpecs track by $index">
-      <div ng-if="customizationArgSpec.schema.type !== 'bool'" class="oppia-interaction-customization-label"><[customizationArgSpec.description]></div>
-      <schema-based-editor ng-class="{'boolean-checkbox': customizationArgSpec.schema.type === 'bool'}"
-                           local-value="StateCustomizationArgsService.displayed[customizationArgSpec.name].value"
-                           schema="customizationArgSpec.schema">
-      </schema-based-editor>
-      <div ng-if="customizationArgSpec.schema.type === 'bool'" class="oppia-interaction-customization-label"><[customizationArgSpec.description]></div>
-      <div class="oppia-interaction-customization-label-container"></div>
-    </div>
-    <div ng-if="getCustomizationArgsWarningMessage()" class="alert alert-danger oppia-units-input-error">
-      <[getCustomizationArgsWarningMessage()]>
-    </div>
-  </div>
-
-  <div ng-if="StateInteractionIdService.displayed && !hasCustomizationArgs && customizationModalReopened">
-    This interaction cannot be modified. To delete it, close this popup and click the 'x' button in the top-right corner.
-  </div>
-</div>
-
-<div class="modal-footer">
-  <button class="btn btn-secondary" ng-click="cancel()" ng-if="!hasCustomizationArgs">Cancel</button>
-  <!-- Do not show this button if an interaction is already selected. -->
-  <button class="btn btn-secondary float-left" ng-click="returnToInteractionSelector()" ng-if="StateInteractionIdService.displayed && !StateInteractionIdService.savedMemento && hasCustomizationArgs">Change Interaction</button>
-
-  <button class="btn btn-secondary" ng-click="cancelWithConfirm()" ng-if="hasCustomizationArgs">Cancel</button>
-  <div class="btn-group"
-       ng-if="hasCustomizationArgs"
-       popover-append-to-body
-       popover-placement="bottom"
-       popover-trigger="<[isSaveInteractionButtonEnabled() ? 'none' : 'mouseenter']>"
-       ng-attr-uib-popover="<[getSaveInteractionButtonTooltip()]>">
-    <button class="btn btn-success protractor-test-save-interaction"
-            ng-click="save()"
-            ng-disabled="!isSaveInteractionButtonEnabled()">
-      Save Interaction
-    </button>
-  </div>
-  <button class="btn btn-success protractor-test-close-no-customization-modal"
-          ng-click="confirm('okay')"
-          ng-if="!hasCustomizationArgs && customizationModalReopened">
-    Okay
-  </button>
-</div>
-<style>
-  .interaction-editor-form .boolean-checkbox {
-    float: left;
-    margin-right: 4px;
-  }
-  .customize-interaction-body-container .oppia-interaction-tile img {
-    border: 1px solid #aaa;
-    border-bottom: none;
-    border-radius: 5px 5px 0 0;
-  }
-  .customize-interaction-body-container ul .active {
-    background-color: #e0f2f1;
-    border-bottom: 2px solid #00609c;
-    color: #00609c;
-  }
-  .customize-interaction-body-container ul :hover:not(.active) {
-    background-color: #e0f2f166;
-    border-bottom: 2px solid #e0f2f166;
-    border-color: #e0f2f166;
-  }
-   .customize-interaction-modal .modal-dialog {
-     margin: 30px auto;
-     max-width: none;
-     width: 45vw;
-  }
-  .customize-interaction-modal .modal-header i {
-    display: none;
-  }
-  .oppia-interaction-header-container .oppia-modal-header {
-    border-bottom: 0;
-    padding-bottom: 0;
-  }
-  .oppia-interaction-header-container .oppia-interaction-customization-label-container {
-    margin-bottom: 15px;
-  }
-  @media screen and (max-width: 768px) {
-    .customize-interaction-modal .modal-content {
-      border: 0;
-      border-radius: 0;
-      margin: 0;
-      min-height: 100vh;
-    }
-    .customize-interaction-modal .modal-dialog {
-      margin: 0;
-      max-width: none;
-      min-height: 100vh;
-      width: 100vw;
-    }
-    .customize-interaction-modal .modal-header-blue {
-      background-color: #00609c;
-    }
-    .customize-interaction-modal .modal-header-blue i {
-      color: #fff;
-      display: block;
-      font-size: 20px;
-    }
-    .customize-interaction-modal .modal-header-blue h3 {
-      color: #fff;
-    }
-  }
-</style>
